requirements.txt
requests==2.31.0
googlesearch-python==1.2.0
urllib3==1.26.16
